update and cleanup scripts, remove duplicate dependency, delete categories, rename to EnvironmentAssumptions
