+++ data-query-tests/pom.xml
@@ -12,18 +12,11 @@
   <packaging>jar</packaging>
   <properties>
     <fhir-resources.version>3.0.15</fhir-resources.version>
-    <junit-platform.version>1.5.2</junit-platform.version>
     <health-apis-ids.version>3.0.0</health-apis-ids.version>
     <selenium.version>3.141.59</selenium.version>
     <sentinel.skipLaunch>false</sentinel.skipLaunch>
   </properties>
   <dependencies>
-    <dependency>
-      <groupId>org.junit.platform</groupId>
-      <artifactId>junit-platform-console-standalone</artifactId>
-      <version>${junit-platform.version}</version>
-      <scope>runtime</scope>
-    </dependency>
     <dependency>
       <groupId>gov.va.api.health</groupId>
       <artifactId>argonaut</artifactId>
