--- data-query-tests/src/main/java/gov/va/api/health/dataquery/tests/categories/InternalBulkFhir.java
@@ -1,3 +1,0 @@
-package gov.va.api.health.dataquery.tests.categories;
-
-public interface InternalBulkFhir {}
