--- data-query-tests/src/main/java/gov/va/api/health/dataquery/tests/categories/LabDataQueryPatient.java
@@ -1,3 +1,0 @@
-package gov.va.api.health.dataquery.tests.categories;
-
-public interface LabDataQueryPatient {}
