--- data-query-tests/src/main/java/gov/va/api/health/dataquery/tests/categories/ProdDataQueryClinician.java
@@ -1,3 +1,0 @@
-package gov.va.api.health.dataquery.tests.categories;
-
-public interface ProdDataQueryClinician {}
