+++ data-query-tests/src/test/java/gov/va/api/health/dataquery/tests/r4/R4PostSearchIT.java
@@ -1,6 +1,6 @@
 package gov.va.api.health.dataquery.tests.r4;
 
-import static gov.va.api.health.dataquery.tests.TestAssumptionUtility.assumeNotLocal;
+import static gov.va.api.health.dataquery.tests.EnvironmentAssumptions.assumeNotLocal;
 
 import gov.va.api.health.dataquery.tests.ResourceVerifier;
 import gov.va.api.health.uscorer4.api.resources.AllergyIntolerance;
